[compression] must call .end() on Inflater/Deflater
When providing an Inflater/Deflater to a *OutputStream,
it is the callers responsibility to cleanup after use.

When created, Inflater/Deflater calls into zlib and allocates a small amount of memory,
which will only get released after .end() is called.
This memory is off-heap, allocated in a C-library, so it will not show up in the JVM,
making for a subtle increase over time.

This fixes memory leak in compression for jose4j.